+++ src/components/CvcSection/CvcSection.tsx
@@ -1,13 +1,25 @@
 import { InputSection } from '../InputSection/InputSection';
 import styles from './CvcSection.module.css';
+import { Dispatch, SetStateAction, useState } from 'react';
 
 type Props = {
   cvc: string;
-  cvcError: string;
-  onChange: (value: string) => void;
+  setCvc: Dispatch<SetStateAction<string>>;
 };
 
-export default function CvcSection({ cvc, cvcError, onChange }: Props) {
+export default function CvcSection({ cvc, setCvc }: Props) {
+  const [cvcError, setCvcError] = useState<string>('');
+
+  const handleCvcChange = (value: string) => {
+    let errorMsg = '';
+    if (!/^[0-9]*$/.test(value)) {
+      errorMsg = '숫자만 입력 가능합니다.';
+    } else if (value !== '' && value.length !== 3) {
+      errorMsg = 'CVC는 3자리여야 합니다.';
+    }
+    setCvc(value);
+    setCvcError(errorMsg);
+  };
   return (
     <div className={styles.sectionContainer}>
       <InputSection.TitleWrapper>
@@ -16,7 +28,7 @@ export default function CvcSection({ cvc, cvcError, onChange }: Props) {
       <InputSection.Label text="CVC" />
       <InputSection.InputWrapper
         numbers={[cvc]}
-        onChange={(_index, value) => onChange(value)}
+        onChange={(_index, value) => handleCvcChange(value)}
         valid={[cvcError === '']}
         placeholders={['123']}
         maxLength={3}
